🚨 engineProcessのundefinedチェックを追加

+++ src/background.ts
@@ -704,28 +704,31 @@ async function restartEngine(engineId: string) {
       runEngine(engineId);
       resolve();
     };
-    engineProcess.once("close", restartEngineOnProcessClosedCallback);
 
-    // treeKillのコールバック関数はコマンドが終了した時に呼ばれます。
-    log.info(
-      `ENGINE ${engineId}: Killing current process (PID=${engineProcess.pid})...`
-    );
-    treeKill(engineProcess.pid, (error) => {
-      // error変数の値がundefined以外であればkillコマンドが失敗したことを意味します。
-      if (error != null) {
-        log.error(`ENGINE ${engineId}: Failed to kill process`);
-        log.error(error);
+    if (engineProcess) {
+      engineProcess.once("close", restartEngineOnProcessClosedCallback);
 
-        // killに失敗したとき、closeイベントが発生せず、once listenerが消費されない
-        // listenerを削除してENGINEの意図しない再起動を防止
-        engineProcess.removeListener(
-          "close",
-          restartEngineOnProcessClosedCallback
-        );
+      // treeKillのコールバック関数はコマンドが終了した時に呼ばれます。
+      log.info(
+        `ENGINE ${engineId}: Killing current process (PID=${engineProcess.pid})...`
+      );
+      treeKill(engineProcess.pid, (error) => {
+        // error変数の値がundefined以外であればkillコマンドが失敗したことを意味します。
+        if (error != null) {
+          log.error(`ENGINE ${engineId}: Failed to kill process`);
+          log.error(error);
 
-        reject();
-      }
-    });
+          // killに失敗したとき、closeイベントが発生せず、once listenerが消費されない
+          // listenerを削除してENGINEの意図しない再起動を防止
+          engineProcess.removeListener(
+            "close",
+            restartEngineOnProcessClosedCallback
+          );
+
+          reject();
+        }
+      });
+    }
   });
 }
